meal-prep/app/whisper_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
openai-whisper
python-multipart==0.0.9
